Move code to lambda function. WIP
